implemented pair_reciprocal for papertrading
